requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==2.2.20
djangorestframework==3.11.2
djangorestframework-jsonapi==2.8.0
inflection==0.5.1
polls==0.1
pytz==2021.3
sqlparse==0.4.2
